+++ src/services/excel-service/properties.ts
